requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
psycopg2-binary
pandas
openai
httpx
asyncio 